requirements.txt
streamlit>=1.49.1
langchain>=0.3.27
langchain-google-genai>=2.1.10
python-dotenv>=1.0.0
